+++ vrp-cli/tests/unit/extensions/solve/config_test.rs
@@ -24,35 +24,45 @@ fn can_read_config() {
     assert_eq!(initial.methods.unwrap().len(), 1);
     assert_eq!(initial.size, Some(1));
 
-    let selection = config.selection.expect("no selection config");
-    assert_eq!(selection.name, "default-naive");
-    assert_eq!(selection.collection.len(), 1);
+    config.selection.expect("no selection config");
+
+    let mutation_config = config.mutation.expect("cannot get mutation");
+    match mutation_config {
+        MutationType::Composite { inners, .. } => {
+            assert_eq!(inners.len(), 3);
+            match inners.first().unwrap() {
+                MutationType::LocalSearch { probability, times, operators: inners } => {
+                    assert_eq!(*probability, 0.05);
+                    assert_eq!(*times, MinMaxConfig { min: 1, max: 2 });
+                    assert_eq!(inners.len(), 4);
+                }
+                _ => unreachable!(),
+            }
+
+            match inners.get(1).unwrap() {
+                MutationType::RuinRecreate { probability, ruins, recreates } => {
+                    assert_eq!(*probability, 1.);
+                    assert_eq!(ruins.len(), 6);
+                    assert_eq!(recreates.len(), 8);
+                }
+                _ => unreachable!(),
+            }
+
+            match inners.last().unwrap() {
+                MutationType::LocalSearch { probability, times, operators: inners } => {
+                    assert_eq!(*probability, 0.01);
+                    assert_eq!(*times, MinMaxConfig { min: 1, max: 2 });
+                    assert_eq!(inners.len(), 4);
+                }
+                _ => unreachable!(),
+            }
+        }
+        _ => unreachable!(),
+    }
 
     let termination = config.termination.expect("no termination config");
     assert_eq!(termination.max_time, Some(300));
     assert_eq!(termination.max_generations, Some(3000));
-
-    let mutation_config = config.mutation.expect("cannot get mutation");
-    assert_eq!(mutation_config.name, "default-composite");
-    assert_eq!(mutation_config.collection.len(), 2);
-
-    match mutation_config.collection.first().unwrap() {
-        MutationType::RuinRecreate { name, ruins, recreates, locals } => {
-            assert_eq!(name, "default-ruin-recreate");
-
-            assert_eq!(ruins.len(), 6);
-            assert_eq!(recreates.len(), 8);
-
-            assert_eq!(locals.pre_ruin.probability, 0.05);
-            assert_eq!(locals.pre_ruin.times, MinMaxConfig { min: 1, max: 2 });
-            assert_eq!(locals.pre_ruin.operators.len(), 4);
-
-            assert_eq!(locals.post_recreate.probability, 0.01);
-            assert_eq!(locals.post_recreate.times, MinMaxConfig { min: 1, max: 2 });
-            assert_eq!(locals.post_recreate.operators.len(), 4);
-        }
-        _ => unreachable!(),
-    }
 }
 
 #[test]