+++ vrp-core/src/solver/selection/mod.rs
@@ -9,5 +9,5 @@ pub use self::naive_selection::NaiveSelection;
 /// A trait which specifies evolution selection behavior.
 pub trait Selection {
     /// Selects parent from population based on refinement process state.
-    fn select_parents(&self, refinement_ctx: &RefinementContext) -> Vec<InsertionContext>;
+    fn select_parents<'a>(&self, refinement_ctx: &'a RefinementContext) -> Vec<&'a InsertionContext>;
 }