+++ CHANGELOG.md
@@ -7,6 +7,11 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - add a new LS method to push tour departure in the future
+- add a Local Search mutation
+
+### Changed
+
+- move Local Search out of Ruin Recreate mutation
 
 
 ## [v1.6.4] - 2020-10-27